requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.7.1
pydantic==2.12.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.34.51
google-generativeai==0.8.3
